fix(endpoint): All dev     - fix Endpoint composers         - fix GET         - add DELETE     - fix Endpoint POST composers: [] operator not supported for strings     - fix Endpoint MasterclassUser:         - add endpoint Delete         - fix voter         - update eventSubscriber         - custom response POST     - clean Forum endpoint     - fix Forum voter     - add endpoint DELETE Forum

+++ app/back/src/Manager/MasterclassUserManager.php
@@ -2,37 +2,63 @@
 
 namespace App\Manager;
 
+use ApiPlatform\Validator\ValidatorInterface;
 use App\Entity\Masterclass;
 use App\Entity\MasterclassUser;
 use App\Entity\User;
+use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
+use Exception;
 
-class MasterclassUserManager
+readonly class MasterclassUserManager
 {
-    public function __construct(private readonly EntityManagerInterface $entityManager, private readonly SectionUserManager $sectionUserManager)
-    {
+    public function __construct(
+        private EntityManagerInterface $entityManager,
+        private SectionUserManager $sectionUserManager,
+        private ValidatorInterface $validator
+    ) {
     }
 
+    /**
+     * @throws Exception
+     */
     public function create(Masterclass $masterclass, User $user, bool $flush = true): MasterclassUser
     {
-        $masterclassUser = new MasterclassUser(['masterclass' => $masterclass, 'user' => $user]);
+        $masterclassUser = $this->entityManager->getRepository(MasterclassUser::class)->findOneBy([
+            'masterclass' => $masterclass, 'user' => $user,
+        ]);
 
-        foreach ($masterclass->getSections() as $section) {
-            $sectionUser = $this->sectionUserManager->create($section, $user, false);
-            $masterclassUser->addSectionUser($sectionUser);
+        if ($masterclassUser instanceof MasterclassUser) {
+            if ($masterclassUser->getDeletedAt() instanceof DateTime) {
+                $masterclassUser->setDeletedAt(null);
+            }
+        } else {
+            $masterclassUser = new MasterclassUser(['masterclass' => $masterclass, 'user' => $user]);
+
+            foreach ($masterclass->getSections() as $section) {
+                $sectionUser = $this->sectionUserManager->create($section, $user, false);
+                $masterclassUser->addSectionUser($sectionUser);
+            }
         }
 
         if ($flush) {
             $this->entityManager->persist($masterclassUser);
+            $this->validator->validate($masterclassUser);
             $this->entityManager->flush();
         }
 
         return $masterclassUser;
     }
 
-    public function delete(MasterclassUser $masterclassUser): void
+    public function delete(Masterclass $masterclass, User $user): void
     {
-        $this->entityManager->remove($masterclassUser);
-        $this->entityManager->flush();
+        $masterclassUser = $this->entityManager->getRepository(MasterclassUser::class)->findOneBy([
+            'masterclass' => $masterclass, 'user' => $user,
+        ]);
+
+        if ($masterclassUser instanceof \App\Entity\MasterclassUser) {
+            $this->entityManager->remove($masterclassUser);
+            $this->entityManager->flush();
+        }
     }
 }
